# inventory-management-system-openapi3_1.yaml
openapi: 3.1.0
info:
  title: Inventory Management System - OpenAPI 3.1
  description: |-
    This is a inventory management system server based on the OpenAPI 3.1 specification. here i am starting with the design first approach in mind! with i can improve the API in general and expose some of the new features in OAS3.
  contact:
    email: 1221344@student.birzeit.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://localhost:9090/ims-api/v1
...
paths:
  /stocks:
    get:
      tags:
        - Stock
      summary: Gets all stocks in the system
      description: Gets all stocks in the system
      operationId: getAllStocks
      responses:
        '200':
          description: The stocks was successfully acquired.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to the stocks
        '404':
          description: Stocks not found.
        '410':
          description: The stocks are no longer available
        '500':
          description: The server encountered an unexpected error while processing the stocks.
        '504':
          description: Unable to process the stocks in time
      security:
        - api_key: []
    post:
      tags: 
        - Stock
      summary: Adds a new stock to the system
      description: Adds a new stock to the system
      operationId: createStock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewStock'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '400':
          description: Invaild stock data provided
        '401':
          description: The client is unauthorized. the stock was not created
        '409':
          description: The stock provided will conflict the resources on the server
        '415':
          description: Content-type of the request is not supported by this server
        '500':
          description: The server encountered an unexpected error while creating the stock
...
  /warehouses:
    get:
      tags:
        - Warehouse
      summary: Gets all warehouses in the system
      description: Gets all warehouses in the system
      operationId: getAllWarehouses
      responses:
        '200':
          description: The warehouses was successfully acquired
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Warehouse'
        '403':
          description: The client doesn't have access to the warehouses
        '404':
          description: there are no warehouses
        '410':
          description: The warehouses are no longer available
        '500':
          description: The server encountered while processing the warehouses
        '504':
          description: Unable to process the warehouses in time
    post:
      tags:
        - Warehouse
      summary: Adds new warehouse to the system
      description: Adds new warehouse to the system
      operationId: createWarehouse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWarehouse'
      responses:
        '201':
          description: The warehouse was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          description: Invaild warehouse data provided
        '401':
          description: The client is unauthorized. the warehouse was not created
        '409':
          description: The warehouse provided will conflict the resources on the server
        '415':
          description: Content-type of the request is not supported by this server
        '500':
          description: The server encountered an unexpected error while creating the warehouse
  /warehouses/{id}:
    get:
      tags:
        - Warehouse
      summary: Gets a warehouse by id
      description: warehouse by id
      operationId: getWarehouseById
      parameters:
        - name: id
          in: path
          description: ID of warehouse to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The warehouse was successfully acquired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '403':
          description: The client doesn't have access to this warehouse
        '404':
          description: there are no warehouse with given id
        '410':
          description: The warehouse with spicified id are no longer available
        '500':
          description: The server encountered an unexpected error while processing the warehouse
        '504':
          description: Unable to process the spicified warehouse in time
    put:
      tags:
        - Warehouse
      summary: Updates a warehouse by id
      description: update warehouse by id
      operationId: updateWarehouseById
      parameters:
        - name: id
          in: path
          description: ID of warehouse to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewWarehouse'
      responses:
        '200':
          description: The warehouse was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          description: Invalid warehouse id
        '401':
          description: The client is unauthorized. the warehouse was not updated
        '404':
          description: Warehouse with spicified id not found
        '409':
          description: The new warehouse data provided will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while updating the warehouse
    delete:
      tags:
        - Warehouse
      summary: Deletes a warehouse by id
      description: Deletes warehouse by id
      operationId: deleteWarehouseById
      parameters:
        - name: id
          in: path
          description: ID of warehouse to return
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The warehouse was successfully deleted
          content:
            application/json:
              schema:
                type: string
                examples: ["The warehouse was successfully deleted"]
        '400':
          description: Invalid warehouse id
        '401':
          description:  The client is unauthorized. the warehouse was not deleted
        '403':
          description:  The client doesn't have access to spicified warehouse. Was unable to deleted
        '404':
          description: Warehouse Not Found
        '500':
          description: The server encountered an unexpected error while deleting the warehouse
  /warehouses/{id}/stocks:
    get:
      tags:
        - Warehouse
      summary: Gets a all stocks in the given warehouse by id
      description: Gets a all stocks in the given warehouse by id
      operationId: getStocksFromWarehouseById
      parameters:
        - name: id
          in: path
          description: ID of warehouse
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stocks was successfully acquired.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '403':
          description: The client doesn't have access to the stocks
        '404':
          description: Stocks not found.
        '410':
          description: The stocks are no longer available
        '500':
          description: The server encountered an unexpected error while processing the stocks.
        '504':
          description: Unable to process the stocks in time
  /warehouses/{id}/linkStock/{stock_id}:
    put:
      tags:
        - Warehouse
      summary: Link a stock to spicific warehouse
      description: Link a stock to spicific warehouse
      operationId: linkStock
      parameters:
        - name: id
          in: path
          description: ID of warehouse to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
        - name: stock_id
          in: path
          description: ID of stock to link
          required: true
          schema:
            type: integer
            format: int64
            examples: [1]
      responses:
        '200':
          description: The stock was successfully linked with the warehouse
          content:
            application/json:
              schema:
                type: string
                examples: [The stock was successfully linked with the warehouse]
        '400':
          description: Invalid warehouse or stock id
        '401':
          description: The client is unauthorized. the warehouse was not linked with the stock
        '404':
          description: Warehouse or Stock with spicified id not found
        '409':
          description: The link will conflict the resources on the server
        '500':
          description: The server encountered an unexpected error while linking the warehouse
...
components:
  schemas:
    Stock:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          examples: [1]
        name:
          type: string
          examples: ["bag"]
        price:
          type: number
          examples: [12.2, 303]
        status:
          type: string
          examples: [available]
          enum:
            - available
            - unavailable
            - pending
        tags:
          type: array
          items:
            type: string
          minItems: 1
          examples: [["reuseable"]]
        availableQuantity:
          type: integer
          format: int64
          examples: [12]
    NewStock:
      type: object
      properties:
        name:
          type: string
          examples: ["bag"]
        price:
          type: number
          examples: [12.2, 303]
        status:
          type: string
          examples: [AVAILABLE]
          enum:
            - AVAILABLE
            - RESERVED
            - OUT_OF_STOCK
            - DAMAGED
            - DISCONTINUED
        tags:
          type: array
          items:
            type: string
          minItems: 1
          examples: [["reuseable"]]
        availableQuantity:
          type: integer
          format: int64
          examples: [12]
    Warehouse:
      type: object
      properties:
        id: 
          type: integer
          format: int64
          examples: [1]
        name: 
          type: string
          examples: [alnoor]
        location:
          type: string
          examples: [Ramallah]
        manager_id: 
          type: integer
          format: int64
          examples: [1]
    NewWarehouse:
      type: object
      properties:
        name: 
          type: string
          examples: [alnoor]
        location:
          type: string
          examples: [Ramallah]
        manager_id: 
          type: integer
          format: int64
          examples: [1]
...